refactor: co-locate zustand stores with their owning features
Move each store out of the central ~/stores directory next to the
component(s) that own it, matching the repo's code-colocation convention:

- recipes-store        -> app/recipes/
- navigation-store     -> components/ (beside navigation-handler/button)
- taste-profile-drawer-store -> components/taste-profile/
- chat-store, chat-drawer-store -> components/chat/ (plurality of consumers)

Same-folder consumers use relative imports; cross-folder consumers use the
~/ alias. Pure move + import-path update, no store API changes. ~/stores
no longer exists.

Closes #546
