+++ src/app/recipes/infinite-recipes.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react'
 import { api } from '~/trpc/react'
 import { useInView } from 'react-intersection-observer'
 import { Recipes } from './recipes'
-import { useRecipesStore } from '~/stores/recipes-store'
+import { useRecipesStore } from './recipes-store'
 
 /** On a desktop the user sees 12 at most. */
 const RECIPES_PER_PAGE_LIMIT = 12
