+++ src/components/auth/sign-up-form.tsx
@@ -10,7 +10,7 @@ import { signIn } from 'next-auth/react'
 import { toast } from '../toast'
 import { signUpSchema, type SignUpSchema } from '~/schemas/sign-up-schema'
 import { api } from '~/trpc/react'
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from '~/components/chat/chat-store'
 import type { MessageWithRecipes } from '~/schemas/chats-schema'
 import { DrawerDialog } from '../drawer-dialog'
 import { UserPlusIcon } from 'lucide-react'
