+++ src/components/chat-panel.tsx
@@ -3,9 +3,9 @@
 import { useEffect } from 'react'
 import { MessageSquareIcon, PlusIcon, XIcon } from 'lucide-react'
 import { Drawer as DrawerPrimitive } from 'vaul'
-import { useChatDrawerStore } from '~/stores/chat-drawer-store'
+import { useChatDrawerStore } from '~/components/chat/chat-drawer-store'
 import type { ChatContext } from '~/schemas/chats-schema'
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from '~/components/chat/chat-store'
 import { Interface } from '~/components/chat/interface'
 import { BottomActiveFilters } from '~/components/chat/bottom-active-filters'
 import { GenerateMessageForm } from '~/components/chat/generate-message-form'
