+++ src/components/chat/extract-tool-invocations.ts
@@ -1,4 +1,4 @@
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from './chat-store'
 import type {
   FullRecipe,
   GeneratedRecipe,
