+++ src/components/chat/generate-message-form.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useTranslations } from '~/hooks/use-translations'
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from './chat-store'
 import { useChat } from '@ai-sdk/react'
 import type { MessageWithRecipes } from '~/schemas/chats-schema'
 import { useChatAI } from '~/hooks/use-chat-ai'
@@ -10,7 +10,7 @@ import { userMessageDTO } from '~/lib/user-message-dto'
 import { useUserId } from '~/hooks/use-user-id'
 import { api } from '~/trpc/react'
 import { selectActiveFilters } from '~/hooks/use-filters-by-user-id'
-import { useChatDrawerStore } from '~/stores/chat-drawer-store'
+import { useChatDrawerStore } from './chat-drawer-store'
 import { Button } from '~/components/button'
 import { Input } from '~/components/ui/input'
 import { BottomBar } from '~/components/bottom-bar'
