+++ src/components/chat/interface.tsx
@@ -5,7 +5,7 @@ import { ScreenLoader } from '~/components/loaders/screen'
 import { ChatWelcome } from './chat-welcome'
 import { AssistantMessageLoader } from '~/components/loaders/assistant-message'
 import { ScrollToBottomProvider } from '~/components/scroll-to-bottom-button'
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from './chat-store'
 import type { MessageWithRecipes } from '~/schemas/chats-schema'
 import { Message } from './message'
 import { AnimatePresence } from '~/components/motion/animate-presence'
