+++ src/components/chat/message.tsx
@@ -4,7 +4,7 @@ import type { MessageWithRecipes } from '~/schemas/chats-schema'
 import { CollapsableRecipe } from './collapsable-recipe'
 import { RecipesToGenerate } from './recipes-to-generate'
 import { useTranslations } from '~/hooks/use-translations'
-import { useChatStore } from '~/stores/chat-store'
+import { useChatStore } from './chat-store'
 import { GenerateStatusAppMessage, ToolResultAppMessage } from './app-message'
 import { Avatar } from './avatar'
 import { FadeIn } from '~/components/motion/fade-in'
